hh/requirements.txt
asgiref==3.5.0
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
coreapi==2.3.3
coreschema==0.0.4
distlib==0.3.4
Django==4.0.3
djangorestframework==3.13.1
drf-yasg==1.20.0
filelock==3.7.0
Flask==2.0.3
fpdf==1.7.2
idna==3.3
inflection==0.5.1
itsdangerous==2.1.1
itypes==1.2.0
Jinja2==3.0.3
loguru==0.6.0
MarkupSafe==2.1.1
packaging==21.3
pika==1.2.0
Pillow==9.1.0
platformdirs==2.5.2
pyparsing==3.0.9
pytz==2022.1
requests==2.27.1
ruamel.yaml==0.17.21
ruamel.yaml.clib==0.2.6
six==1.16.0
sqlparse==0.4.2
tzdata==2022.1
uritemplate==4.1.1
urllib3==1.26.9
virtualenv==20.14.1
Werkzeug==2.0.3
win32-setctime==1.1.0
